comm: Solve TODO check if sent

+++ src/comm/stream.sh
@@ -15,8 +15,9 @@ do
 
 		echo "Sending file: $FILE..."
 		sshpass -p "$PWRD" scp $TO_STREAM/$FILE $LOCAL:$REMOTE_PATH
-		# TODO: remove sent files
-		rm $TO_STREAM/$FILE
-
+		RETVAL=$?
+		if [ $RETVAL -eq 0 ] ; then
+			rm $TO_STREAM/$FILE
+		fi
 	done
 done
